reopen file when truncating, use int64 offset
Make the offset int64-flavored (this doesn't quite fix the
underlying risk of the offset wrapping, but we are a ways from
that mattering still). Also, when truncating a file, reopen
it. This is because Truncate doesn't change the write position
offset, but you can't Seek on a file opened with O_APPEND.

+++ roaring/roaring.go
@@ -1637,7 +1637,7 @@ func (b *Bitmap) unmarshalPilosaRoaring(data []byte) error {
 		var opr op
 		if err := opr.UnmarshalBinary(buf); err != nil {
 			// return error with position so file can be trimmed.
-			return FileCorruptionError{err: err, TrimTo: goodOffset}
+			return FileCorruptionError{err: err, TrimTo: int64(goodOffset)}
 		}
 
 		opr.apply(b)
@@ -1664,7 +1664,7 @@ func (b *Bitmap) unmarshalPilosaRoaring(data []byte) error {
 
 type FileCorruptionError struct {
 	err    error
-	TrimTo int
+	TrimTo int64
 }
 
 func (f FileCorruptionError) Error() string {
